Refactoring: add HeaderTest class, remove testLogOut there remove methods and fields from DashboardPage to BasePage
