+++ src/test/java/ui/MainPageTest.java
@@ -13,16 +13,6 @@ public class MainPageTest extends BaseTest {
     }
 
     @Test
-    public void testLogOut() {
-        String actualLoginMessageText = new DashboardPage(getDriver())
-                .hoverOnUserProfileButton()
-                .clickLogOutButton()
-                .getLoginMessageText();
-
-        Assert.assertEquals(actualLoginMessageText, "You are now logged out.");
-    }
-
-    @Test
     public void testHelpPanelIsOpened(){
        boolean actualResult = new DashboardPage(getDriver())
                 .helpButtonClick()
