+++ src/test/java/ui/model/base/BasePage.java
@@ -1,9 +1,21 @@
 package ui.model.base;
 
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.interactions.Actions;
+import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
+import ui.model.DashboardPage;
+import ui.model.LoginPage;
 
 public abstract class BasePage {
+
+    @FindBy(xpath = "//a[@href='http://localhost:8000/wp-admin/profile.php']")
+    private WebElement userProfileButton;
+
+    @FindBy(xpath = "//a[text()='Log Out']")
+    private WebElement logOutButton;
+
     private final WebDriver driver;
     protected WebDriver getDriver() {
         return driver;
@@ -14,5 +26,17 @@ public abstract class BasePage {
         PageFactory.initElements(driver, this);
     }
 
+    public DashboardPage hoverOnUserProfileButton() {
+        new Actions(getDriver())
+                .moveToElement(userProfileButton)
+                .perform();
+
+        return new DashboardPage(getDriver());
+    }
+
+    public LoginPage clickLogOutButton() {
+        logOutButton.click();
+        return new LoginPage(getDriver());
+    }
 
 }
