+++ src/test/java/ui/runner/BaseTest.java
@@ -2,18 +2,23 @@ package ui.runner;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.support.FindBy;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 
 import java.net.URL;
 
 public abstract class BaseTest {
+
     private static WebDriver driver;
 
     public static WebDriver getDriver() {
         return driver;
     }
+
+
 
     @BeforeMethod
     public static void setUp() {
